moved plotpeople into gui.py
